+++ core/php/AbeilleMonitor.php
@@ -44,7 +44,7 @@
 
         /* Write to fifo */
         $queue = msg_get_queue($abQueues['parserToMon']['id']);
-        msg_send($queue, 1, $msg);
+        msg_send($queue, 1, json_encode($msg), false);
     }
 
     /* Called from AbeilleParser when short address has changed (device announce) */
@@ -61,7 +61,7 @@
 
         /* Write to fifo */
         $queue = msg_get_queue($abQueues['parserToMon']['id']);
-        msg_send($queue, 1, $msg);
+        msg_send($queue, 1, json_encode($msg), false);
     }
 
     /* Called from AbeilleCmd to add a message to monitor.
@@ -87,7 +87,7 @@
 
         /* Write to fifo */
         $queue = msg_get_queue($abQueues['cmdToMon']['id']);
-        msg_send($queue, 1, $msg);
+        msg_send($queue, 1, json_encode($msg), false);
     }
 
     /* Called to inform 'AbeilleCmd' to change monitoring address. */
@@ -103,7 +103,7 @@
 
         /* Write to fifo */
         $queue = msg_get_queue($abQueues['monToCmd']['id']);
-        msg_send($queue, 1, $msg);
+        msg_send($queue, 1, json_encode($msg), false);
     }
 
     /* Shutdown function */
@@ -189,18 +189,19 @@
                 if (($statCmdQueue['msg_qnum'] != 0) && ($statParserQueue['msg_qnum'] != 0)) {
                     /* Select queue with oldest message */
                     if ($statCmdQueue['msg_stime'] < $statParserQueue['msg_stime'])
-                        msg_receive($queueCmdToMon, 0, $msgType, $max_msg_size, $msg, true, MSG_IPC_NOWAIT);
+                        msg_receive($queueCmdToMon, 0, $msgType, $max_msg_size, $msgJson, false, MSG_IPC_NOWAIT);
                     else {
-                        msg_receive($queueParserToMon, 0, $msgType, $max_msg_size, $msg, true, MSG_IPC_NOWAIT);
+                        msg_receive($queueParserToMon, 0, $msgType, $max_msg_size, $msgJson, false, MSG_IPC_NOWAIT);
                         $msgSent = false;
                     }
                 } else if ($statCmdQueue['msg_qnum'] != 0)
-                    msg_receive($queueCmdToMon, 0, $msgType, $max_msg_size, $msg, true, MSG_IPC_NOWAIT);
+                    msg_receive($queueCmdToMon, 0, $msgType, $max_msg_size, $msgJson, false, MSG_IPC_NOWAIT);
                 else {
-                    msg_receive($queueParserToMon, 0, $msgType, $max_msg_size, $msg, true, MSG_IPC_NOWAIT);
+                    msg_receive($queueParserToMon, 0, $msgType, $max_msg_size, $msgJson, false, MSG_IPC_NOWAIT);
                     $msgSent = false;
                 }
 
+                $msg = json_decode($msgJson, true);
                 if ($msg['type'] == 'x2mon') {
                     /* Log message */
                     if ($msgSent == true)