+++ desktop/php/AbeilleEq-Main-Icon.php
@@ -10,7 +10,7 @@
 <div class="form-group">
     <label class="col-sm-3 control-label">{{Icone}}</label>
     <div class="col-sm-3">
-        <select id="sel_icon" class="form-control eqLogicAttr" data-l1key="configuration" data-l2key="icone">
+        <select id="sel_icon" class="form-control eqLogicAttr" data-l1key="configuration" data-l2key="ab::icon">
             <!-- <option value="Abeille">{{Abeille}}</option> -->
             <option value="Ruche">{{Ruche}}</option>
             <?php